EqualizerView: Abstract out visualizations to allow others
* Random generated EQ visualization is now known as `SampledVisualization`
