+++ app/src/main/res/layout/playlist_track_entity.xml
@@ -26,7 +26,7 @@
             android:visibility="visible"
             app:layout_constraintHorizontal_chainStyle="packed"/>
 
-    <dev.bmcreations.guacamole.ui.widgets.EqualizerView
+    <dev.bmcreations.guacamole.ui.widgets.visualization.EqualizerView
             android:id="@+id/equalizer"
             android:layout_width="wrap_content"
             android:layout_height="40dp"
